DI java stereotypes

+++ di-java-stereotypes/src/main/java/eu/balev/student/repository/StudentRepository.java
@@ -1,11 +1,10 @@
 package eu.balev.student.repository;
 
 import eu.balev.student.model.Student;
-
 import java.util.List;
 
 public interface StudentRepository {
     List<Student> getAllStudents();
 
-    int count();
+    long count();
 }
